requirements.txt
telethon
python-telegram-bot==13.7
